+++ src/components/Outlet.ts
@@ -1,18 +1,12 @@
 import { LitElement, RootPart, html, render } from 'lit';
-import { createElement, type ComponentType } from 'react';
-import { createRoot } from 'react-dom/client';
+import { createRoot, Root } from 'react-dom/client';
+import { type ReactElement } from 'react';
+import { type TemplateResult } from 'lit';
 
 interface RenderOption {
   id?: string;
   force?: boolean;
-}
-
-interface RenderElementOption extends RenderOption {
-  element: typeof LitElement | string;
-}
-
-interface RenderComponentOption extends RenderOption {
-  component: ComponentType;
+  content: HTMLElement | ReactElement | TemplateResult<1>;
 }
 
 /**
@@ -20,12 +14,9 @@ interface RenderComponentOption extends RenderOption {
  * 라우트 설정에 따라 LitElement 또는 React 컴포넌트를 렌더링합니다.
  */
 export class Outlet extends LitElement {
-
   private routeId?: string;
   private container?: HTMLDivElement;
-
-  private reactRoot?: ReturnType<typeof createRoot>;
-  private litRoot?: RootPart;
+  private content?: Root | RootPart;
 
   /** 쉐도우를 사용하지 않고, 직접 렌더링합니다. */
   protected createRenderRoot() {
@@ -37,71 +28,57 @@ export class Outlet extends LitElement {
   }
 
   /**
-   * 기존 렌더링된 DOM을 제거하고, LitElement를 삽입합니다.
-   * - 이전 ID와 동일하고, 강제 렌더링이 아닌 경우, 이전 DOM을 반환합니다.
+   * render 함수의 결과를 렌더링합니다.
+   * - HTMLElement, ReactElement, TemplateResult를 모두 처리할 수 있습니다.
    */
-  public async renderElement({ id, element, force }: RenderElementOption) {
-    if (this.routeId === id && !force) {
-      return this.container?.firstElementChild;
-    }
-
-    this.routeId = id;
-    this.clear();
-    const template = typeof element === 'string'
-      ? document.createElement(element)
-      : element.prototype instanceof LitElement
-      ? new element()
-      : undefined;
-    if (!template || !this.container) {
-      throw new Error('DOM이 초기화되지 않았습니다.');
-    }
-    this.litRoot = render(html`${template}`, this.container);
-
-    this.requestUpdate();
-    await this.updateComplete;
-    return template;
-  }
-
-  /**
-   * 기존 렌더링된 DOM을 제거하고, React 컴포넌트를 삽입합니다.
-   * - 이전 ID와 동일하고, 강제 렌더링이 아닌 경우, 이전 DOM을 반환합니다.
-   */
-  public async renderComponent({ id, component, force }: RenderComponentOption) {
-    if (this.routeId === id && !force) {
+  public async renderContent({ id, content, force }: RenderOption) {
+    if (this.routeId === id && force === false && this.container) {
       return this.container;
     }
 
     this.routeId = id;
     this.clear();
-    const template = createElement(component);
-    if (!template || !this.container) {
+    
+    if (!this.container) {
       throw new Error('DOM이 초기화되지 않았습니다.');
     }
-    this.reactRoot = createRoot(this.container);
-    this.reactRoot.render(template);
+
+    if (content instanceof HTMLElement) {
+      // HTMLElement인 경우 직접 추가
+      this.container.appendChild(content);
+    } else if ("_$litType$" in content) {
+      // TemplateResult인 경우
+      this.content = render(content as TemplateResult<1>, this.container);
+    } else if ('$$typeof' in content) {
+      // ReactElement인 경우
+      this.content = createRoot(this.container);
+      this.content.render(content as ReactElement);
+    } else {
+      throw new Error('not supported content type for Outlet rendering.');
+    }
 
     this.requestUpdate();
     await this.updateComplete;
-    return this.container;
+    return this.container!;
   }
 
   /**
    * 기존 DOM을 라이프 사이클에 맞게 제거합니다.
    */
   public clear() {
-    // React DOM이 렌더링된 경우
-    if(this.reactRoot) {
-      this.reactRoot.unmount();
-      this.reactRoot = undefined;
+    if (this.content) {
+      // React DOM이 렌더링된 경우
+      if ('unmount' in this.content) {
+        this.content.unmount();
+      }
+      // LitElement가 렌더링된 경우
+      if ('setConnected' in this.content) {
+        this.content.setConnected(false);
+      }
+      this.content = undefined;
     }
-    // LitElement가 렌더링된 경우
-    if (this.litRoot) {
-      this.litRoot.setConnected(false);
-      this.litRoot = undefined;
-    }
-    // Container 초기화
+    // Container 초기화 - display: contents로 레이아웃에 영향을 주지 않음
     this.container = document.createElement('div');
-    this.container.style.width = '100%';
-    this.container.style.height = '100%';
+    this.container.style.display = 'contents';
   }
 }