+++ src/types/index.ts
@@ -1,5 +1,5 @@
 export type { RouteConfig } from './RouteConfig';
 export type { RouteInfo } from './RouteInfo';
 export type { RouterConfig } from './RouterConfig';
-export type { RouteError } from './RouteError';
+export { RouteError, NotFoundRouteError } from './RouteError';
 export * from './RouteEvent';